Created EasyAi that randomly chooses a column though will prevent you from winning if 3 in a row and similarly try to win if it has 3 in a row.
Also changed what the Board class is, simply put it should just be a
board with insert/remove disc. All methods that were in there have been
moved to the GameSystem class, since that should handle all the rule
checking, win/draw checking. I also generalised the win check code
though pretty sure it still works completely; made it more scalable, say
for a 9x19 board of connect 6?

Created an interface between the AI and GameSystem to check whether a
move is legal, able to check win. Similarly, also changed how the AI
receives the board and what it can access because of this, so Iai
interface is also changed.